--- src/components/10-atoms/input-text/react/story.jsx
@@ -1,113 +1,0 @@
-/* global document */
-import React, { PureComponent } from 'react';
-import ReactDOM from 'react-dom';
-import changelog from '../CHANGELOG.md';
-import readme from '../README.md';
-import AXAInputText from './AXAInputText';
-
-export default {
-  title: 'Examples/Input Text/React',
-  parameters: {
-    readme,
-    changelog,
-    controls: { disabled: false },
-    options: { showPanel: true },
-  },
-};
-
-const typeOptions = {
-  text: 'text',
-  email: 'email',
-  password: 'password',
-};
-
-export const Story = ({
-  label,
-  name,
-  refId,
-  placeholder,
-  value,
-  currency,
-  error,
-  info,
-  checkMark,
-  disabled,
-  required,
-  invalid,
-  defaultValue,
-  types,
-  maxLength,
-  counter,
-  counterMax,
-  pattern,
-  inputmode,
-  autofocus,
-}) => {
-  class InputText extends PureComponent {
-    constructor() {
-      super();
-
-      this.state = {
-        value,
-      };
-    }
-
-    render() {
-      return (
-        <AXAInputText
-          refId={refId}
-          name={name}
-          label={label}
-          placeholder={placeholder}
-          checkMark={checkMark}
-          disabled={disabled}
-          required={required}
-          invalid={invalid}
-          defaultValue={defaultValue}
-          value={this.state.value}
-          type={types}
-          error={error}
-          info={info}
-          maxLength={maxLength}
-          counter={counter}
-          counterMax={counterMax}
-          pattern={pattern}
-          inputmode={inputmode}
-          autofocus={autofocus}
-          currency={currency}
-          onChange={evt => this.setState({ value: evt.target.value })}
-        />
-      );
-    }
-  }
-
-  const wrapper = document.createElement('div');
-  ReactDOM.render(<InputText />, wrapper);
-
-  return wrapper;
-};
-
-Story.args = {
-  label: '',
-  name: '',
-  refId: '',
-  placeholder: '',
-  value: '',
-  error: '',
-  info: '',
-  checkMark: false,
-  disabled: false,
-  required: false,
-  invalid: false,
-  types: 'text',
-  maxLength: 50,
-  counter: 'Still ##counter## characters left',
-  counterMax: 'Over character limit!',
-  pattern: '',
-  inputmode: '',
-  autofocus: false,
-};
-
-Story.argTypes = {
-  types: { name: 'type', control: { type: 'radio', options: typeOptions } },
-};
